+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Release v3.21.13
 ### Update
 - Update UMKA token address [#1880](https://github.com/kvhnuke/etherwallet/pull/1880)
+- [Network] linting css & walletDecryptDrtv.html [#1878](https://github.com/kvhnuke/etherwallet/pull/1880)
 - Minor grammatical fixes [#1873](https://github.com/kvhnuke/etherwallet/pull/1873)
 - Update CLO explorer address [#1872](https://github.com/kvhnuke/etherwallet/pull/1872)
 - Fix network tab css [#1870](https://github.com/kvhnuke/etherwallet/pull/1870)