+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Release v3.21.13
 ### Update
+- Update VEN to VET []()
 - Update UMKA token address [#1880](https://github.com/kvhnuke/etherwallet/pull/1880)
 - [Network] linting css & walletDecryptDrtv.html [#1878](https://github.com/kvhnuke/etherwallet/pull/1880)
 - Minor grammatical fixes [#1873](https://github.com/kvhnuke/etherwallet/pull/1873)
@@ -7,6 +8,9 @@
 - Fix network tab css [#1870](https://github.com/kvhnuke/etherwallet/pull/1870)
 
 ### New
+- Add CARD token [#1884](https://github.com/kvhnuke/etherwallet/pull/1884)
+- Add ORBS token [#1883](https://github.com/kvhnuke/etherwallet/pull/1883)
+- Add C8 token [#1881](https://github.com/kvhnuke/etherwallet/pull/1881)
 - Add COIN token [#1880](https://github.com/kvhnuke/etherwallet/pull/1880)
 - YAP Node [#1876](https://github.com/kvhnuke/etherwallet/pull/1876)
 - Add EGT, WYS, IOTX, BTCL, GZR and FYX tokens [#1874](https://github.com/kvhnuke/etherwallet/pull/1874)